PickCertificationChallenges returns challenges
