+++ api/lib/domain/usecases/retrieve-last-or-create-certification-course.js
@@ -1,7 +1,6 @@
 const CertificationCourse = require('../models/CertificationCourse');
 const Assessment = require('../models/Assessment');
 const { UserNotAuthorizedToCertifyError, NotFoundError } = require('../errors');
-const _ = require('lodash');
 
 module.exports = async function retrieveLastOrCreateCertificationCourse({
   domainTransaction,
@@ -65,14 +64,16 @@ async function _startNewCertification({
     throw new UserNotAuthorizedToCertifyError();
   }
 
+  // Le mot Profile = ???? >
+  // - PlacementProfile (AssResult),
+  // - Scorecards (pas calculé pareil ? KEs),
+  // - TargetProfile,
+  // - CampaignProfile (contient un placementProfile)
   // certficationProfile = CertificationProfile.from(placementProfile);
 
-  // FIXME : Ce n'est pas la responsabilité du placementProfile que d'avoir les challenges
-  const userCompetences = await placementProfileService.pickCertificationChallenges(placementProfile);
-  const challenges = _.flatMap(userCompetences, (userCompetence) => userCompetence.challenges);
-  // FIXME: userCompetences, est-ce les mêmes partout ? sachant que celles qu'on a fabriqué ci-dessus prennent en compte les modifications de jury
+  // FIXME : Est-ce que pickCertificationChallenges est une fonction de placementProfileService ?
+  const challenges = await placementProfileService.pickCertificationChallenges(placementProfile);
   const newCertificationChallenges = certificationChallengesService.generateCertificationChallenges(challenges);
-  // certificationCourseChallenges = identifierLesChallenges(certificationProfile); // $$$
 
   const existingCertificationCourse = await certificationCourseRepository.findOneCertificationCourseByUserIdAndSessionId({
     userId,
